LoglessPairHMM arrays implementation
Instead of using [readbases]x[haplotypebases] matrices to store interim
values for the insert, delete, and match HMM states, we perform the same
dynamic programming using three generations of 1-dimensional arrays.

The order in which read/haplotype base intersections are encountered is
changed. We proceed as if passing over the diagonals of a read x haplotype
matrix.

This way, values in the arrays do not depend on each other, only on the
previous array generations. Updates are thus amenable to parallelization.

Currently no haplotype caching is performed.

Passes the smallTest and the mediumTest