Remove .hpp implementations Move update method inside diagonals and now unused zero() method Move diagonals to it's own class with all relevant operations Move constants to it's own class with all relevant operations Make padded_reads a private member of PairhmmImpl Move make test to top-level makefile Add 'make test' to Travis automated ci
Although I fully appreciate the effort of reducing compilation time,
the hpp split is not reasonable. The hpp is not a compilable standalone
unit and coding in it is a nightmare (without syntax checking). We need
a better standard. For now, let's keep it all in the .h

Fixes #31
Fixes #27
Fixes #26